quiz-api/requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycodestyle==2.13.0
PyJWT==2.5.0
tomli==2.2.1
Werkzeug==3.1.3
